+++ src/core/HLE/XAPI/Xapi.h
